Distributed_Event_Management_System/back-end/event-registration/requirements.txt
flask
boto3
flask-socketio
flask-cors
python-dotenv
